refactor: made new structure of the project with modules
