+++ src/js/index.js
@@ -1,238 +1,20 @@
-import { availableHotelsSlider, homesSlider, reviewsSlider } from '../modules/slider';
-import getHomes from './services/getHomes';
+import { homesSlider, reviewsSlider } from './modules/slider';
+import { closeAdv } from './modules/closeAdv';
+import { menuBurger } from './modules/menuBurger';
+import filter from './modules/form/filter';
+import homesCards from './modules/homesCards';
+import form from './modules/form/form';
 
-//  filter
-
-const selectWrapper = document.querySelector('.select__wrapper');
-const filterWrapper = document.querySelector('.filter__wrapper');
-const inputAdults = document.getElementById('adults');
-const inputChildren = document.getElementById('children');
-const inputRooms = document.getElementById('rooms');
-const counterInputAdults = document.querySelector('.item__counter--adults');
-const counterInputChildren = document.querySelector('.item__counter--children');
-const additional = document.querySelector('.filter__additional-content');
-const counterInputRooms = document.querySelector('.item__counter--rooms');
-const counterBtnsAdults = document.querySelectorAll('.filter__button--adults');
-const counterBtnsChildren = document.querySelectorAll('.filter__button--children');
-const counterBtnsRooms = document.querySelectorAll('.filter__button--rooms');
-
-const showFilter = () => {
-  filterWrapper.classList.remove('hide');
-  filterWrapper.classList.add('show');
-};
-
-const hideFilter = () => {
-  filterWrapper.classList.remove('show');
-  filterWrapper.classList.add('hide');
-};
-
-selectWrapper.addEventListener('click', showFilter);
-
-window.addEventListener('click', (e) => {
-  if (!e.target.closest('.select__wrapper') && !e.target.closest('.filter__wrapper')) {
-    hideFilter();
-  }
-});
-
-const counterAdults = 1;
-const counterChildren = 0;
-const counterRooms = 1;
-let childrenNums = [];
-
-const changeInputCounter = (
-  buttonsType,
-  counterType,
-  counterInputType,
-  inputType,
-  minCount,
-  maxCount,
-  name,
-  value,
-  addValue,
-) => {
-  buttonsType.forEach((btn, index) => {
-    buttonsType[0].disabled = true;
-    btn.addEventListener('click', (e) => {
-      e.preventDefault();
-      if (index === 0 && counterType > minCount) {
-        counterType--;
-        if (e.target.classList.contains('filter__button--children--dec')) {
-          document.querySelector('.filter__additional-select').remove();
-        }
-      } else if (index === 1 && counterType < maxCount) {
-        counterType++;
-        if (e.target.classList.contains('filter__button--children--inc')) {
-          childrenNums.push(counterType);
-          childrenNums.map((i) => {
-            const selectWrapperInner = document.createElement('div');
-            selectWrapperInner.classList.add('select--wrapper');
-            selectWrapperInner.innerHTML = `<select class="filter__additional-select" id=${i}>
-                                        ${new Array(17)
-                                          .fill(0)
-                                          .map(
-                                            (item, indexOption) =>
-                                              `<option value=${indexOption}>${indexOption} years old</option>`,
-                                          )}
-                                     </select>`;
-            additional.appendChild(selectWrapperInner);
-          });
-          childrenNums = [];
-        }
-      }
-      buttonsType[0].disabled = counterType === value;
-      if (counterType === addValue) {
-        additional.classList.remove('hide');
-      } else if (counterType === 0) additional.classList.add('hide');
-      counterInputType.value = counterType;
-      inputType.value = `${counterType} ${name}`;
-    });
-  });
-};
-
-changeInputCounter(counterBtnsAdults, counterAdults, counterInputAdults, inputAdults, 1, 30, 'Adults', 1);
-changeInputCounter(counterBtnsChildren, counterChildren, counterInputChildren, inputChildren, 0, 10, 'Children', 0, 1);
-changeInputCounter(counterBtnsRooms, counterRooms, counterInputRooms, inputRooms, 1, 30, 'Rooms', 1);
-
-// homes cards
-
-const homesSliderWrapper = document.querySelector('.homes__slider-wrapper');
 const url = 'https://if-student-api.onrender.com/api/hotels';
-
-const getCards = (cards) => {
-  for (let j = 0; j < cards.length; j++) {
-    for (let i = 0; i < cards.length - 1; i++) {
-      let a = cards[i];
-      let b = cards[i + 1];
-      if (a.name > b.name) {
-        const temp = a;
-        a = b;
-        b = temp;
-      }
-      cards[i] = a;
-      cards[i + 1] = b;
-    }
-    console.log(cards);
-  }
-
-  cards.forEach(({ city, country, imageUrl, name }) => {
-    const homesSlide = document.createElement('div');
-    homesSlide.classList.add('homes__slide', 'swiper-slide');
-    homesSlide.innerHTML = `
-                  <div class="homes__slide-image">
-                    <img
-                      class="homes__slide-image-item"
-                      src=${imageUrl}
-                      alt=""
-                    />
-                  </div>
-                  <a href="#" class="homes__slide-title">${name}</a>
-                  <div class="homes__slide-subtitle">
-                    ${city}, ${country}
-                  </div>
-  `;
-    homesSliderWrapper.append(homesSlide);
-  });
-};
+//  filter
+filter();
+// homes cards
+homesCards(url);
 homesSlider();
-
-if (!sessionStorage.getItem('hotels')) {
-  getHomes(url).then((data) => {
-    getCards(data);
-    sessionStorage.setItem('hotels', JSON.stringify(data));
-  });
-} else {
-  getCards(JSON.parse(sessionStorage.getItem('hotels')));
-}
-
 // form
-
-const availableHotelSliderWrapper = document.querySelector('.available-hotels__slider-wrapper');
-const availableHotelsSection = document.querySelector('.available-hotels');
-const destInput = document.getElementById('destination');
-const form = document.querySelector('.content__form');
-const preloader = document.querySelector('.available-hotels__preloader');
-const emptyData = document.querySelector('.empty-data');
-const adultsInput = document.getElementById('adults');
-const roomsInput = document.getElementById('rooms');
-
-const getAvailableHomes = (hotels) => {
-  availableHotelSliderWrapper.innerHTML = '';
-  hotels.forEach(({ city, country, imageUrl, name }) => {
-    const availableHotelsSlide = document.createElement('div');
-    availableHotelsSlide.classList.add('available-hotels__slide');
-    availableHotelsSlide.innerHTML = `
-                  <div class="available-hotels__slide-image">
-                    <img
-                      class="available-hotels__slide-image-item"
-                      src=${imageUrl}
-                      alt=""
-                    />
-                  </div>
-                  <a href="#" class="available-hotels__slide-title">${name}</a>
-                  <div class="available-hotels__slide-subtitle">
-                    ${city}, ${country}
-                  </div>
-  `;
-    availableHotelSliderWrapper.append(availableHotelsSlide);
-    availableHotelsSlide.classList.add('swiper-slide');
-  });
-};
-availableHotelsSlider();
-
-const showAvailableHotels = () => {
-  const urlAvailable = new URL(url);
-  const ages = [];
-
-  document.querySelectorAll('select').forEach((select) => {
-    ages.push(select.options.selectedIndex);
-  });
-  const childrenStr = ages.join(',');
-
-  urlAvailable.searchParams.append('search', `${destInput.value}`);
-  urlAvailable.searchParams.append('adults', `${adultsInput.value.replace('Adults', '')}`);
-  urlAvailable.searchParams.append('children', `${childrenStr}`);
-  urlAvailable.searchParams.append('rooms', `${roomsInput.value.replace('Rooms', '')}`);
-  if (destInput.value !== '') {
-    preloader.classList.remove('hide');
-    getHomes(urlAvailable).then((data) => {
-      if (data.length) {
-        preloader.classList.add('hide');
-        emptyData.classList.add('hide');
-        getAvailableHomes(data);
-        availableHotelsSection.classList.remove('hide');
-      } else {
-        preloader.classList.add('hide');
-        emptyData.classList.remove('hide');
-        availableHotelsSection.classList.add('hide');
-      }
-    });
-  }
-};
-
-form.addEventListener('submit', (e) => {
-  e.preventDefault();
-  showAvailableHotels();
-});
-
+form(url);
 // close adv
-
-const advert = document.querySelector('.advertisement');
-const closeAdvBtn = document.querySelector('.advertisement__content-close');
-
-const closeAdv = () => {
-  closeAdvBtn.addEventListener('click', (e) => {
-    e.preventDefault();
-    advert.classList.add('hide');
-  });
-};
 closeAdv();
 //  menu burger
-
-const burgerBtn = document.querySelector('.menu__burger');
-const burgerContent = document.querySelector('.menu-burger-content');
-
-burgerBtn.addEventListener('click', () => {
-  burgerContent.classList.toggle('show');
-  burgerBtn.classList.toggle('active');
-});
+menuBurger();
 reviewsSlider();
