+++ src/js/modules/slider.js
