--- src/js/services/getHomes.js
@@ -1,9 +1,0 @@
-const getHomes = async (url) => {
-  const res = await fetch(url);
-  if (!res.ok) {
-    throw new Error(`Could not fetch${url}, error - ${res.status}`);
-  }
-  return res.json();
-};
-
-export default getHomes;
